Route properly for authenticated user.
